+++ app/services/download_epub_cover.rb
@@ -1,7 +1,0 @@
-class DownloadEpubCover
-  def self.call journal, url
-    journal.epub_cover.download! url
-    journal.save
-    journal
-  end
-end