+++ client/app/pods/components/edit-journal-epub-css/component.js
@@ -1,5 +1,0 @@
-import EditJournalCSS from 'tahi/pods/components/edit-journal-css/component';
-
-export default EditJournalCSS.extend({
-  type: 'epub'
-});